Make systemctl try to fallback to upstart.
This is needed to handle the transition phase after systemd
has been installed but before the system has been rebooted.
During this period, the legacy upstart init is still in use
and is responsible for shutting the system down. Thus, make
sure that systemctl (i.e. the new init userland) is able to
tell the old init to do things (reboot, primarily).